refactor: every enum declares its smallest underlying type; flatten slnx folders
Audited all 61 enum declarations under src/ and tests/. Four were not on
the smallest integral type that holds their members: Keys (ReplayFrame),
ComparisonOperator, UpdateCheckOutcome and DotnetLaunchMode had no
explicit type (defaulting to int) and now declare byte; ClientFlags
declared uint but its highest flag is 1 << 13, so it is now ushort (the
sibling LastFmFlags keeps uint for its 1 << 22 member). Every other enum
already declared the minimal type. No member values changed, so no wire
or storage encoding changes -- every explicit cast still produces the
same numeric value.

Basil.slnx: solution folders cannot nest in the slnx format; the
/Sources/Protocol/ folder introduced by D1 is now written flat as a
sibling of /Sources/ instead of inside it.

Verified together with the preceding commit (same build/test runs).

+++ src/Basil.Domain/Scores/ClientFlags.cs
@@ -9,7 +9,7 @@ namespace Basil.Domain.Scores;
 ///     or used solely for telemetry/data collection.
 /// </remarks>
 [Flags]
-public enum ClientFlags : uint
+public enum ClientFlags : ushort
 {
 	/// <summary>
 	///     The client triggered no anticheat flags.
